renaming properties and general cleanup
Attempting to make our property names more descriptive and less cryptic.  For instance, a variable which is a raw string to be passed to a regex compile (or similar) has a _re_s suffix.  If the regex has already been compiled, the suffix is just _re.
Also, I'm taking some first steps toward making our version look and act a little more like the original php (where appropriate, of course).