Move _glyph_defaults to Textile.glyph_definitions.
Having them (back) in the Textile class allows changing them in a subclass,
for example to change quote marks to ones used in a local language.
Renaming to glyph_definitions, because the dictionary is used directly
rather than being a default.

Added a simple test case for sublassing Textile and changing a glyph.